requirements.txt
# split mp3

# brew install libav --with-libvorbis --with-sdl --with-theora
# brew install taglib
pytaglib

pydub
minio
